src_analysis/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
ca-certificates=2022.07.19=h06a4308_0
certifi=2021.5.30=py36h06a4308_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
expat=2.4.4=h295c915_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.63.1=h5a9c865_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=he6710b0_3
intel-openmp=2020.2=254
jpeg=9e=h7f8727e_0
kiwisolver=1.3.1=py36h2531618_0
lcms2=2.12=h3be6417_0
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h2818925_1
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
matplotlib-base=3.3.2=py36h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.3=h5eee18b_3
numpy=1.15.4=py36h7e9f1db_0
numpy-base=1.15.4=py36hde5b4d6_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.0.2u=h7b6447c_0
pandas=1.1.3=py36he6710b0_0
pcre=8.45=h295c915_0
pillow=8.3.1=py36h5aabda8_0
pip=21.2.2=py36h06a4308_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.5=hc3d631a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
qt=5.9.6=h8703b6f_2
readline=7.0=h7b6447c_5
scipy=1.1.0=py36h7c811a0_2
setuptools=58.0.4=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
tk=8.6.12=h1ccaba5_0
tornado=6.1=py36h27cfd23_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
